+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Added italian translation
+
 ## 1.0.2
 * Lives left are now shown as a title instead of the chat
 * Fixed "lives remaining" showing up in creative/spectator and hardcore