+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 1.1.0
 * Added max life config. Also added a new command argument to cap the maximum lives you can get
 * Players' lives lost and remaining are now announced to chat on death
 * Player now actually respawns (on the position he died) on last life and gets a message